found hidden issue and other minor things
found an issue with the path splitting which adds some random tile in front of the spawn tile in the path list, which was breaking some other stuff i was trying to make work in mapgenerator. that works now, this way we don't have a horrendous number of path tiles.

this should help me later with figuring out sprite drawing of the paths.

i still want to figure out what the core issue is in TileSetGenerator that was causing this